# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HireLoop Campus Recruitment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/analyze-resume:
    post:
      operationId: analyzeResume
      tags: [ai]
      summary: AI resume analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeResumeBody"
      responses:
        "200":
          description: Resume analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeAnalysis"
  /ai/mock-interview/start:
    post:
      operationId: startMockInterview
      tags: [ai]
      summary: Start a mock interview session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartInterviewBody"
      responses:
        "200":
          description: Interview session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewSession"
  /ai/mock-interview/{sessionId}/answer:
    post:
      operationId: submitInterviewAnswer
      tags: [ai]
      summary: Submit answer and get next question
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAnswerBody"
      responses:
        "200":
          description: AI feedback and next question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerFeedback"
  /ai/mock-interview/{sessionId}/summary:
    get:
      operationId: getInterviewSummary
      tags: [ai]
      summary: Get interview session summary
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewSummary"
...
  /admin/students:
    get:
      operationId: listStudents
      tags: [admin]
      summary: List all students (admin)
      parameters:
        - in: query
          name: branch
          schema:
            type: string
        - in: query
          name: placed
          schema:
            type: boolean
      responses:
        "200":
          description: Students list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentProfile"
...
  /admin/recruiters/{recruiterId}/approve:
    put:
      operationId: approveRecruiter
      tags: [admin]
      summary: Approve or reject a recruiter (admin)
      parameters:
        - in: path
          name: recruiterId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveRecruiterBody"
      responses:
        "200":
          description: Updated recruiter status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecruiterProfile"
...
components:
  schemas:
...
    StudentProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        branch:
          type: string
        batch:
          type: string
        cgpa:
          type: number
        skills:
          type: array
          items:
            type: string
        bio:
          type: string
        linkedinUrl:
          type: string
        githubUrl:
          type: string
        placementStatus:
          type: string
          enum: [unplaced, placed, internship]
        placedCompany:
          type: string
        packageLpa:
          type: number
        resumeScore:
          type: integer
        profileCompleteness:
          type: integer
        totalApplications:
          type: integer
        shortlisted:
          type: integer
        offers:
          type: integer
      required: [id, userId, name, email, branch, batch, cgpa, skills, placementStatus, profileCompleteness]
...
    AnalyzeResumeBody:
      type: object
      properties:
        jobDescription:
          type: string
      required: []
    ResumeAnalysis:
      type: object
      properties:
        atsScore:
          type: integer
        matchPercentage:
          type: integer
        keywordGaps:
          type: array
          items:
            type: string
        suggestions:
          type: array
          items:
            type: string
        strengths:
          type: array
          items:
            type: string
        sectionScores:
          type: object
          properties:
            summary:
              type: integer
            experience:
              type: integer
            skills:
              type: integer
            projects:
              type: integer
            education:
              type: integer
          required: [summary, experience, skills, projects, education]
      required: [atsScore, matchPercentage, keywordGaps, suggestions, strengths, sectionScores]
    StartInterviewBody:
      type: object
      properties:
        role:
          type: string
        level:
          type: string
          enum: [fresher, junior, mid]
      required: [role]
    InterviewSession:
      type: object
      properties:
        sessionId:
          type: string
        role:
          type: string
        level:
          type: string
        status:
          type: string
          enum: [active, completed]
        currentQuestion:
          type: string
        questionNumber:
          type: integer
        totalQuestions:
          type: integer
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        overallScore:
          type: integer
      required: [sessionId, role, level, status, questionNumber, totalQuestions, startedAt]
    SubmitAnswerBody:
      type: object
      properties:
        answer:
          type: string
      required: [answer]
    AnswerFeedback:
      type: object
      properties:
        feedback:
          type: string
        score:
          type: integer
        nextQuestion:
          type: string
        isComplete:
          type: boolean
        improvementTip:
          type: string
      required: [feedback, score, isComplete]
    InterviewSummary:
      type: object
      properties:
        sessionId:
          type: string
        role:
          type: string
        communicationScore:
          type: integer
        technicalScore:
          type: integer
        overallScore:
          type: integer
        strengths:
          type: array
          items:
            type: string
        improvements:
          type: array
          items:
            type: string
        questionAnswers:
          type: array
          items:
            type: object
            properties:
              question:
                type: string
              answer:
                type: string
              score:
                type: integer
              feedback:
                type: string
            required: [question, answer, score, feedback]
      required: [sessionId, role, communicationScore, technicalScore, overallScore, strengths, improvements, questionAnswers]
...
    RecruiterProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        company:
          type: string
        designation:
          type: string
        phone:
          type: string
        website:
          type: string
        isApproved:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, userId, name, email, company, isApproved, createdAt]
...
    ApproveRecruiterBody:
      type: object
      properties:
        isApproved:
          type: boolean
      required: [isApproved]